+++ pykwalify/rule.py
@@ -2,7 +2,7 @@
 
 """ pyKwalify - rule.py """
 
-# python std lib
+# python stdlib
 import logging
 import os
 import re
@@ -65,13 +65,13 @@ class Rule(object):
         return "Rule: {}".format(str(self._schema_str))
 
     def init(self, schema, path):
-        log.debug("Init schema: {}".format(schema))
+        log.debug(u"Init schema: {}".format(schema))
 
         include = schema.get("include", None)
 
         # Check if this item is a include, overwrite schema with include schema and continue to parse
         if include:
-            log.debug("Found include tag...")
+            log.debug(u"Found include tag...")
             self._include_name = include
             return
 
@@ -88,10 +88,10 @@ class Rule(object):
                     t = "map"
                     self.init_type_value(t, rule, path)
                 else:
-                    raise RuleError("key 'type' not found in schema rule : {}".format(path))
+                    raise RuleError(u"key 'type' not found in schema rule : {}".format(path))
             else:
                 if not isinstance(schema["type"], str):
-                    raise RuleError("key 'type' in schema rule is not a string type : {}".format(path))
+                    raise RuleError(u"key 'type' in schema rule is not a string type : {}".format(path))
 
                 self._type = schema["type"]
 
@@ -129,10 +129,10 @@ class Rule(object):
             if k in func_mapping:
                 func_mapping[k](v, rule, path)
             elif k.startswith("schema;"):
-                log.debug("Found schema tag...")
-                raise RuleError("Schema is only allowed on top level of schema file...")
+                log.debug(u"Found schema tag...")
+                raise RuleError(u"Schema is only allowed on top level of schema file...")
             else:
-                raise RuleError("Unknown key: {} found : {}".format(k, path))
+                raise RuleError(u"Unknown key: {} found : {}".format(k, path))
 
         self.check_conflicts(schema, rule, path)
 
@@ -140,7 +140,7 @@ class Rule(object):
         """
         """
         if not isinstance(v, str):
-            raise RuleError("value for func keyword must be a string : {} : {}".format(v, path))
+            raise RuleError(u"value for func keyword must be a string : {} : {}".format(v, path))
 
         self._func = v
 
@@ -148,33 +148,33 @@ class Rule(object):
         """
         """
         if not isinstance(v, list):
-            raise RuleError("Extension defenition should be a list : {} : {}".format(v, path))
+            raise RuleError(u"Extension defenition should be a list : {} : {}".format(v, path))
 
         # TODO: Add limitation that this keyword can only be used at the top level of the file
 
         self._extensions = v
 
     def init_matching_rule(self, v, rule, path):
-        log.debug("Init matching-rule: {}".format(path))
-        log.debug("{} {}".format(v, rule))
+        log.debug(u"Init matching-rule: {}".format(path))
+        log.debug(u"{} {}".format(v, rule))
 
         # Verify that the provided rule is part of one of the allowed one
         allowed = ["any"]
         # ["none", "one", "all"] Is currently awaiting proper implementation
         if v not in allowed:
-            raise RuleError("Specefied rule in key : {} is not part of allowed rule set : {}".format(v, allowed))
+            raise RuleError(u"Specefied rule in key : {} is not part of allowed rule set : {}".format(v, allowed))
         else:
             self._matching_rule = v
 
     def init_allow_empty_map(self, v, rule, path):
-        log.debug("Init allow empty value: {}".format(path))
-        log.debug("Type: {} : {}".format(v, rule))
+        log.debug(u"Init allow empty value: {}".format(path))
+        log.debug(u"Type: {} : {}".format(v, rule))
 
         self._allowempty_map = v
 
     def init_type_value(self, v, rule, path):
-        log.debug("Init type value : {}".format(path))
-        log.debug("Type: {} {}".format(v, rule))
+        log.debug(u"Init type value : {}".format(path))
+        log.debug(u"Type: {} {}".format(v, rule))
 
         if v is None:
             v = DEFAULT_TYPE
@@ -183,102 +183,102 @@ class Rule(object):
         self._type_class = type_class(v)
 
         if not is_builtin_type(self._type):
-            raise RuleError("type.unknown : {} : {}".format(self._type, path))
+            raise RuleError(u"type.unknown : {} : {}".format(self._type, path))
 
     def init_matching(self, v, rule, path):
-        log.debug("Init matching rule : {}".format(path))
+        log.debug(u"Init matching rule : {}".format(path))
 
         valid_values = ["any", "all", "*"]
 
         if str(v) not in valid_values:
-            raise RuleError("matching value: {} is not one of {}".format(str(v), valid_values))
+            raise RuleError(u"matching value: {} is not one of {}".format(str(v), valid_values))
 
         self._matching = str(v)
 
     def init_name_value(self, v, rule, path):
-        log.debug("Init name value : {}".format(path))
+        log.debug(u"Init name value : {}".format(path))
 
         self._name = str(v)
 
     def init_desc_value(self, v, rule, path):
-        log.debug("Init descr value : {}".format(path))
+        log.debug(u"Init descr value : {}".format(path))
 
         self._desc = str(v)
 
     def init_required_value(self, v, rule, path):
-        log.debug("Init required value : {}".format(path))
+        log.debug(u"Init required value : {}".format(path))
 
         if not isinstance(v, bool):
-            raise RuleError("required.notbool : {} : {}".format(v, path))
+            raise RuleError(u"required.notbool : {} : {}".format(v, path))
         self._required = v
 
     def init_pattern_value(self, v, rule, path):
-        log.debug("Init pattern value : {}".format(path))
+        log.debug(u"Init pattern value : {}".format(path))
 
         if not isinstance(v, str):
-            raise RuleError("pattern.notstr : {} : {}".format(v, path))
+            raise RuleError(u"pattern.notstr : {} : {}".format(v, path))
 
         self._pattern = v
 
         if self._schema_str["type"] == "map":
-            raise RuleError("map.pattern : pattern not allowed inside map : {} : {}".format(v, path))
+            raise RuleError(u"map.pattern : pattern not allowed inside map : {} : {}".format(v, path))
 
         # TODO: Some form of validation of the regexp? it exists in the source
 
         try:
             self._pattern_regexp = re.compile(self._pattern)
         except Exception:
-            raise RuleError("pattern.syntaxerr : {} --> {} : {}".format(self._pattern_regexp, self._pattern_regexp, path))
+            raise RuleError(u"pattern.syntaxerr : {} --> {} : {}".format(self._pattern_regexp, self._pattern_regexp, path))
 
     def init_enum_value(self, v, rule, path):
-        log.debug("Init enum value : {}".format(path))
+        log.debug(u"Init enum value : {}".format(path))
 
         if not isinstance(v, list):
-            raise RuleError("enum.notseq")
+            raise RuleError(u"enum.notseq")
         self._enum = v
 
         if is_collection_type(self._type):
-            raise RuleError("enum.notscalar")
+            raise RuleError(u"enum.notscalar")
 
         lookup = set()
         for item in v:
             if not isinstance(item, self._type_class):
-                raise RuleError("enum.type.unmatch : {} --> {} : {}".format(item, self._type_class, path))
+                raise RuleError(u"enum.type.unmatch : {} --> {} : {}".format(item, self._type_class, path))
 
             if item in lookup:
-                raise RuleError("enum.duplicate : {} : {}".format(item, path))
+                raise RuleError(u"enum.duplicate : {} : {}".format(item, path))
 
             lookup.add(item)
 
     def init_assert_value(self, v, rule, path):
-        log.debug("Init assert value : {}".format(path))
+        log.debug(u"Init assert value : {}".format(path))
 
         if not isinstance(v, str):
-            raise RuleError("assert.notstr : {}".format(path))
+            raise RuleError(u"assert.notstr : {}".format(path))
 
         self._assert = v
 
-        raise RuleError("assert.NYI-Error : {}".format(path))
+        raise RuleError(u"assert.NYI-Error : {}".format(path))
 
     def init_range_value(self, v, rule, path):
-        log.debug("Init range value : {}".format(path))
+        log.debug(u"Init range value : {}".format(path))
 
         if not isinstance(v, dict):
-            raise RuleError("range.notmap : {} : {}".format(v, path))
+            raise RuleError(u"range.notmap : {} : {}".format(v, path))
         if self._type not in ["str", "int", "map", "seq"]:
-            raise RuleError("range.not-supported-type : {} : {}".format(self._type, path))
+            raise RuleError(u"range.not-supported-type : {} : {}".format(self._type, path))
 
         self._range = v  # dict that should contain min, max, min-ex, max-ex keys
 
         # This should validate that only min, max, min-ex, max-ex exists in the dict
         for k, v in self._range.items():
             if k not in ["max", "min", "max-ex", "min-ex"]:
-                raise RuleError("range.undefined key : {} : {}".format(k, path))
+                raise RuleError(u"range.undefined key : {} : {}".format(k, path))
 
         if "max" in self._range and "max-ex" in self._range:
-            raise RuleError("range.twomax : {}".format(path))
+            raise RuleError(u"range.twomax : {}".format(path))
         if "min" in self._range and "min-ex" in self._range:
-            raise RuleError("range.twomin : {}".format(path))
+            raise RuleError(u"range.twomin : {}".format(path))
 
         max = self._range.get("max", None)
         min = self._range.get("min", None)
@@ -286,67 +286,67 @@ class Rule(object):
         min_ex = self._range.get("min-ex", None)
 
         if max is not None and not is_int(max) or is_bool(max):
-            raise RuleError("range.max.notint : {} : {}".format(max, path))
+            raise RuleError(u"range.max.notint : {} : {}".format(max, path))
 
         if min is not None and not is_int(min) or is_bool(min):
-            raise RuleError("range.min.notint : {} : {}".format(min, path))
+            raise RuleError(u"range.min.notint : {} : {}".format(min, path))
 
         if max_ex is not None and not is_int(max_ex) or is_bool(max_ex):
-            raise RuleError("range.max_ex.notint : {} : {}".format(max_ex, path))
+            raise RuleError(u"range.max_ex.notint : {} : {}".format(max_ex, path))
 
         if min_ex is not None and not is_int(min_ex) or is_bool(min_ex):
-            raise RuleError("range.min_ex.notint : {} : {}".format(min_ex, path))
+            raise RuleError(u"range.min_ex.notint : {} : {}".format(min_ex, path))
 
         if max is not None:
             if min is not None and max < min:
-                raise RuleError("range.maxltmin : {} < {} : {}".format(max, min, path))
+                raise RuleError(u"range.maxltmin : {} < {} : {}".format(max, min, path))
             elif min_ex is not None and max <= min_ex:
-                raise RuleError("range.maxleminex : {} <= {} : {}".format(max, min_ex, path))
+                raise RuleError(u"range.maxleminex : {} <= {} : {}".format(max, min_ex, path))
         elif max_ex is not None:
             if min is not None and max_ex < min:
-                raise RuleError("range.maxexlemiin : {} < {} : {}".format(max_ex, min, path))
+                raise RuleError(u"range.maxexlemiin : {} < {} : {}".format(max_ex, min, path))
             elif min_ex is not None and max_ex <= min_ex:
-                raise RuleError("range.maxexleminex : {} <= {} : {}".format(max_ex, min_ex, path))
+                raise RuleError(u"range.maxexleminex : {} <= {} : {}".format(max_ex, min_ex, path))
 
     def init_ident_value(self, v, rule, path):
-        log.debug("Init ident value : {}".format(path))
+        log.debug(u"Init ident value : {}".format(path))
 
         if v is None or isinstance(v, bool):
-            raise RuleError("ident.notbool : {} : {}".format(v, path))
+            raise RuleError(u"ident.notbool : {} : {}".format(v, path))
 
         self._ident = bool(v)
         self._required = True
 
         if is_collection_type(self._type):
-            raise RuleError("ident.notscalar : {} : {}".format(self._type, path))
+            raise RuleError(u"ident.notscalar : {} : {}".format(self._type, path))
         if path == "":
-            raise RuleError("ident.onroot")
+            raise RuleError(u"ident.onroot")
         if self._parent is None or not self._parent._type == "map":
-            raise RuleError("ident.notmap : {}".format(path))
+            raise RuleError(u"ident.notmap : {}".format(path))
 
     def init_unique_value(self, v, rule, path):
-        log.debug("Init unique value : {}".format(path))
+        log.debug(u"Init unique value : {}".format(path))
 
         if not isinstance(v, bool):
-            raise RuleError("unique.notbool : {} : {}".format(v, path))
+            raise RuleError(u"unique.notbool : {} : {}".format(v, path))
 
         self._unique = v
 
         if is_collection_type(self._type):
-            raise RuleError("unique.notscalar : {} : {}".format(self._type, path))
+            raise RuleError(u"unique.notscalar : {} : {}".format(self._type, path))
         if path == "":
-            raise RuleError("unique.onroot")
+            raise RuleError(u"unique.onroot")
 
     def init_sequence_value(self, v, rule, path):
-        log.debug("Init sequence value : {}".format(path))
+        log.debug(u"Init sequence value : {}".format(path))
 
         if v is not None and not isinstance(v, list):
-            raise RuleError("sequence.notseq : {} : {}".format(v, path))
+            raise RuleError(u"sequence.notseq : {} : {}".format(v, path))
 
         self._sequence = v
 
         if self._sequence is None or len(self._sequence) == 0:
-            raise RuleError("sequence.noelem : {} : {}".format(self._sequence, path))
+            raise RuleError(u"sequence.noelem : {} : {}".format(self._sequence, path))
 
         tmp_seq = []
 
@@ -354,7 +354,7 @@ class Rule(object):
             elem = e or {}
 
             rule = Rule(None, self)
-            rule.init(elem, "{}/sequence/{}".format(path, i))
+            rule.init(elem, u"{}/sequence/{}".format(path, i))
 
             tmp_seq.append(rule)
 
@@ -365,15 +365,15 @@ class Rule(object):
     def init_mapping_value(self, v, rule, path):
         # Check for duplicate use of 'map' and 'mapping'
         if self._mapping:
-            raise RuleError("mapping.multiple-use : {}".format(path))
+            raise RuleError(u"mapping.multiple-use : {}".format(path))
 
-        log.debug("Init mapping value : {}".format(path))
+        log.debug(u"Init mapping value : {}".format(path))
 
         if v is not None and not isinstance(v, dict):
-            raise RuleError("mapping.notmap : {} : {}".format(v, path))
+            raise RuleError(u"mapping.notmap : {} : {}".format(v, path))
 
         if v is None or len(v) == 0:
-            raise RuleError("mapping.noelem : {} : {}".format(v, path))
+            raise RuleError(u"mapping.noelem : {} : {}".format(v, path))
 
         self._mapping = {}
         self._regex_mappings = []
@@ -384,66 +384,66 @@ class Rule(object):
 
             # Check if this is a regex rule. Handle specially
             if k.startswith("regex;") or k.startswith("re;"):
-                log.debug("Found regex map rule")
+                log.debug(u"Found regex map rule")
                 regex = k.split(";", 1)
                 if len(regex) != 2:
-                    raise RuleError("Malformed regex key : {}".format(k))
+                    raise RuleError(u"Malformed regex key : {}".format(k))
                 else:
                     regex = regex[1]
                     try:
                         re.compile(regex)
                     except Exception as e:
-                        raise RuleError("Unable to compile regex '{}' '{}'".format(regex, e))
+                        raise RuleError(u"Unable to compile regex '{}' '{}'".format(regex, e))
 
                     regex_rule = Rule(None, self)
-                    regex_rule.init(v, "{}/mapping;regex/{}".format(path, regex[1:-1]))
+                    regex_rule.init(v, u"{}/mapping;regex/{}".format(path, regex[1:-1]))
                     regex_rule._map_regex_rule = regex[1:-1]
                     self._regex_mappings.append(regex_rule)
                     self._mapping[k] = regex_rule
             else:
                 rule = Rule(None, self)
-                rule.init(v, "{}/mapping/{}".format(path, k))
+                rule.init(v, u"{}/mapping/{}".format(path, k))
                 self._mapping[k] = rule
 
         return rule
 
     def init_default_value(self, v, rule, path):
-        log.debug("Init default value : {}".format(path))
+        log.debug(u"Init default value : {}".format(path))
         self._default = v
 
         if is_collection_type(self._type):
-            raise RuleError("default.notscalar : {} : {} : {}".format(rule, path, v))
+            raise RuleError(u"default.notscalar : {} : {} : {}".format(rule, path, v))
 
         if self._type == "map" or self._type == "seq":
-            raise RuleError("default.notscalar : {} : {} : {}".format(rule, os.path.dirname(path), v))
+            raise RuleError(u"default.notscalar : {} : {} : {}".format(rule, os.path.dirname(path), v))
 
         if not isinstance(v, self._type_class):
-            raise RuleError("default.type.unmatch : {} --> {} : {}".format(v, self._type_class, path))
+            raise RuleError(u"default.type.unmatch : {} --> {} : {}".format(v, self._type_class, path))
 
     def check_conflicts(self, schema, rule, path):
-        log.debug("Checking for conflicts : {}".format(path))
+        log.debug(u"Checking for conflicts : {}".format(path))
 
         if self._type == "seq":
             if all([sa not in schema for sa in sequence_aliases]):
-                raise SchemaConflict("seq.nosequence")
+                raise SchemaConflict(u"seq.nosequence")
             if self._enum is not None:
-                raise SchemaConflict("seq.conflict :: enum: {}".format(path))
+                raise SchemaConflict(u"seq.conflict :: enum: {}".format(path))
             if self._pattern is not None:
-                raise SchemaConflict("seq.conflict :: pattern: {}".format(path))
+                raise SchemaConflict(u"seq.conflict :: pattern: {}".format(path))
             if self._mapping is not None:
-                raise SchemaConflict("seq.conflict :: mapping: {}".format(path))
+                raise SchemaConflict(u"seq.conflict :: mapping: {}".format(path))
         elif self._type == "map":
             if all([ma not in schema for ma in mapping_aliases]) and not self._allowempty_map:
-                raise SchemaConflict("map.nomapping")
+                raise SchemaConflict(u"map.nomapping")
             if self._enum is not None:
-                raise SchemaConflict("map.conflict :: enum:")
+                raise SchemaConflict(u"map.conflict :: enum:")
             if self._sequence is not None:
-                raise SchemaConflict("map.conflict :: mapping: {}".format(path))
+                raise SchemaConflict(u"map.conflict :: mapping: {}".format(path))
         else:
             if self._sequence is not None:
-                raise SchemaConflict("scalar.conflict :: sequence: {}".format(path))
+                raise SchemaConflict(u"scalar.conflict :: sequence: {}".format(path))
             if self._mapping is not None:
-                raise SchemaConflict("scalar.conflict :: mapping: {}".format(path))
+                raise SchemaConflict(u"scalar.conflict :: mapping: {}".format(path))
             if self._enum is not None:
                 if self._range is not None:
-                    raise SchemaConflict("enum.conflict :: range: {}".format(path))
+                    raise SchemaConflict(u"enum.conflict :: range: {}".format(path))